requirements.txt
apache-airflow==2.10.4
apache-airflow-providers-amazon==9.2.0
apache-airflow-providers-snowflake==5.5.0
dbt-core==1.9.2
dbt-snowflake==1.9.1
pandas
apache-airflow-providers-http
